deleted unnecessary components, moved components from 'components' directory into 'pages' directory
